[decompiler] Fuse `let X; X = e;` back into `let X = e;`
`hoist_declarations` lifts an arm-scope `let X = e` to a top-level
`Declare([X])` plus an arm-local `Assign([X], e)`. When those two end up
adjacent in the same `Seq` (with matching single targets), fuse them
back into a `LetBind([X], e)` so the output reads as the idiomatic
single declaration with initializer. Anything mismatching (different
targets, multi-target, non-adjacent) is left alone.
