+++ external-crates/move/crates/move-decompiler/src/refinement/remove_trailing_return.rs
@@ -4,8 +4,8 @@
 use crate::ast::Exp;
 
 // Elide a bare `return` (no operands) when it appears in tail position of the function
-// body. Recurses through `Seq` (last element), `IfElse` (both arms), and `Switch` (all
-// cases). Does NOT recurse through loops/whiles — `return` there exits the function,
+// body. Recurs through `Seq` (last element), `IfElse` (both arms), and `Switch` (all
+// cases). Does NOT recur through loops/whiles — `return` there exits the function,
 // not the loop, so it is not interchangeable with falling through.
 pub fn refine(exp: &mut Exp) -> bool {
     elide_tail(exp)
