+++ external-crates/move/crates/move-decompiler/src/structuring/mod.rs
@@ -4,7 +4,10 @@
 pub(crate) mod ast;
 pub(crate) mod dom_tree;
 pub(crate) mod graph;
+pub(crate) mod hoist_declarations;
 pub(crate) mod term_reconstruction;
+
+pub use hoist_declarations::hoist_declarations;
 
 use crate::{
     config::{self, print_heading},
