+++ external-crates/move/crates/move-decompiler/tests/bytecode/output/0xb24efcc5dcf03cf4b1fb0e2155206b4e2c5b008da29ff4025a0db241e4578976/dbke.move
@@ -1,6 +1,3 @@
----
-source: crates/move-decompiler/tests/tests.rs
----
 // Auto-generated by Move decompiler
 
 module 0xb24efcc5dcf03cf4b1fb0e2155206b4e2c5b008da29ff4025a0db241e4578976::dbke;
@@ -265,3 +262,4 @@ fun vbac<T,T>(l0: &0x2c8d603bc51326b8c13cef9dd07031a408a48dddb541963357661df5d32
         }
     }
 }
+
